Renamed model_analysis to insilico
