+++ openretina/insilico/stimulus_optimization/objective.py
