+++ openretina/insilico/stimulus_optimization/optimization_stopper.py
