+++ openretina/insilico/stimulus_optimization/regularizer.py
